NMS-14541 Bridge Topology Discovery, Move BridgeTopologyException
Re reworked static methods
Removed BrbidgeTopologyException from bridgeTopologyService
