+++ features/enlinkd/daemon/src/main/java/org/opennms/netmgt/enlinkd/EnhancedLinkd.java
@@ -42,10 +42,9 @@ import org.opennms.netmgt.config.EnhancedLinkdConfig;
 import org.opennms.netmgt.config.datacollection.SnmpCollection;
 import org.opennms.netmgt.daemon.AbstractServiceDaemon;
 import org.opennms.netmgt.enlinkd.api.ReloadableTopologyDaemon;
+import org.opennms.netmgt.enlinkd.common.Discovery;
 import org.opennms.netmgt.enlinkd.common.NodeCollector;
 import org.opennms.netmgt.enlinkd.common.TopologyUpdater;
-import org.opennms.netmgt.enlinkd.common.Discovery;
-import org.opennms.netmgt.enlinkd.service.api.BridgeTopologyException;
 import org.opennms.netmgt.enlinkd.service.api.BridgeTopologyService;
 import org.opennms.netmgt.enlinkd.service.api.CdpTopologyService;
 import org.opennms.netmgt.enlinkd.service.api.IpNetToMediaTopologyService;
@@ -499,11 +498,7 @@ public class EnhancedLinkd extends AbstractServiceDaemon implements ReloadableTo
                         nodeid);
         unscheduleNodeCollection(nodeid);
 
-        try {
-            m_bridgeTopologyService.delete(nodeid);
-        } catch (BridgeTopologyException e) {
-            LOG.error("deleteNode: {}", e.getMessage());
-        }
+        m_bridgeTopologyService.delete(nodeid);
         m_cdpTopologyService.delete(nodeid);
         m_isisTopologyService.delete(nodeid);
         m_lldpTopologyService.delete(nodeid);
