+++ features/enlinkd/service/api/src/main/java/org/opennms/netmgt/enlinkd/service/api/SharedSegment.java
@@ -28,179 +28,17 @@
 
 package org.opennms.netmgt.enlinkd.service.api;
 
-import java.util.ArrayList;
 import java.util.Date;
-import java.util.HashMap;
 import java.util.HashSet;
-import java.util.List;
-import java.util.Map;
 import java.util.Set;
 
-import org.opennms.netmgt.enlinkd.model.BridgeBridgeLink;
-import org.opennms.netmgt.enlinkd.model.BridgeMacLink;
-import org.opennms.netmgt.enlinkd.model.BridgeMacLink.BridgeMacLinkType;
-import org.opennms.netmgt.model.OnmsNode;
-
 public class SharedSegment implements Topology{
-    
-    public static List<BridgeBridgeLink> getBridgeBridgeLinks(SharedSegment segment) throws BridgeTopologyException {
-        return create(segment.getDesignatedPort(), segment.getBridgePortsOnSegment());
-    }
 
-    public static List<BridgeMacLink> getBridgeMacLinks(SharedSegment segment) throws BridgeTopologyException {
-        return BridgeTopologyService.create(segment.getDesignatedPort(), segment.getMacsOnSegment(), BridgeMacLinkType.BRIDGE_LINK);
-    }
-    
-    public static SharedSegment create(BridgeMacLink link) throws BridgeTopologyException {
-        SharedSegment segment = new SharedSegment();
-        segment.getBridgePortsOnSegment().add(getFromBridgeMacLink(link));
-        segment.getMacsOnSegment().add(link.getMacAddress());
-        segment.setDesignatedBridge(link.getNode().getId());
-        segment.setCreateTime(link.getBridgeMacLinkCreateTime());
-        segment.setLastPollTime(link.getBridgeMacLinkLastPollTime());
-        return segment;
-    }
-
-    public static SharedSegment create(BridgeBridgeLink link) throws BridgeTopologyException {
-        SharedSegment segment = new SharedSegment();
-        segment.getBridgePortsOnSegment().add(getFromBridgeBridgeLink(link));
-        segment.getBridgePortsOnSegment().add(getFromDesignatedBridgeBridgeLink(link));
-        segment.setDesignatedBridge(link.getDesignatedNode().getId());
-        segment.setCreateTime(link.getBridgeBridgeLinkCreateTime());
-        segment.setLastPollTime(link.getBridgeBridgeLinkLastPollTime());
-        return segment;
-    }
-        
-    public static void merge(BroadcastDomain domain,
-        SharedSegment upsegment, 
-        Map<BridgePortWithMacs, Set<BridgePortWithMacs>> splitted, 
-        Set<String> macsonsegment, 
-        BridgePort rootport,
-        Set<BridgePortWithMacs> throughset) {
-        
-        splitted.keySet().forEach(designated -> {
-            Set<BridgePortWithMacs> ports = splitted.get(designated);
-            SharedSegment splitsegment = new SharedSegment();
-            splitsegment.getBridgePortsOnSegment().add(designated.getPort());
-            splitsegment.setDesignatedBridge(designated.getPort().getNodeId());
-            Set<String> macs = new HashSet<>(designated.getMacs());
-            ports.forEach(bft ->
-            {
-                macs.retainAll(bft.getMacs());
-                domain.cleanForwarders(bft.getPort().getNodeId());
-                upsegment.getBridgePortsOnSegment().remove(bft.getPort());
-                splitsegment.getBridgePortsOnSegment().add(bft.getPort());
-            });
-            splitsegment.getMacsOnSegment().addAll(macs);
-            domain.getSharedSegments().add(splitsegment);
-            domain.cleanForwarders(macs);
-        });
-        
-        //Add macs from forwarders
-        Map<String, Integer> forfpmacs = new HashMap<>();
-        upsegment.getBridgePortsOnSegment().forEach(port ->
-        {
-            domain.getForwarders(port.getNodeId()).stream().filter(forward -> forward.getPort().equals(port)).
-            forEach( forward ->
-                    forward.getMacs().forEach(mac -> {
-                        int itemsfound=1;
-                        if (forfpmacs.containsKey(mac)) {
-                            itemsfound = forfpmacs.get(mac);
-                            itemsfound++;
-                        }
-                        forfpmacs.put(mac, itemsfound);
-                    }));
-            
-            Set<String> clearmacs = new HashSet<>();
-            forfpmacs.keySet().forEach(mac -> {
-                if (forfpmacs.get(mac) == upsegment.getBridgePortsOnSegment().size()) {
-                    upsegment.getMacsOnSegment().add(mac);
-                    clearmacs.add(mac);
-                }
-            });
-            domain.cleanForwarders(clearmacs);
-            
-        });
-
-        upsegment.getBridgePortsOnSegment().add(rootport);
-        upsegment.getMacsOnSegment().retainAll(macsonsegment);
-        domain.cleanForwarders(upsegment.getMacsOnSegment());
-
-        throughset.forEach(bft -> SharedSegment.createAndAddToBroadcastDomain(domain,
-                                                                                       bft));
-    }
-
-    public static void createAndAddToBroadcastDomain(BroadcastDomain domain, BridgePortWithMacs bft) {
-        SharedSegment segment = new SharedSegment();
-        segment.getBridgePortsOnSegment().add(bft.getPort());
-        segment.getMacsOnSegment().addAll(bft.getMacs());
-        segment.setDesignatedBridge(bft.getPort().getNodeId());
-        domain.getSharedSegments().add(segment);
-        domain.cleanForwarders(bft.getMacs());
-    }
-        
-    public static SharedSegment create() {
-        return new SharedSegment();
-                
-    }
-    
     private Integer m_designatedBridgeId;
     private final Set<String> m_macsOnSegment = new HashSet<>();
     private final Set<BridgePort> m_portsOnSegment = new HashSet<>();
     private Date m_createTime;
     private Date m_lastPollTime;
-
-    public static BridgePort getFromBridgeMacLink(BridgeMacLink link) {
-        BridgePort bp = new BridgePort();
-        bp.setNodeId(link.getNode().getId());
-        bp.setBridgePort(link.getBridgePort());
-        bp.setBridgePortIfIndex(link.getBridgePortIfIndex());
-        bp.setVlan(link.getVlan());
-        return bp;
-    }
-
-    public static BridgePort getFromBridgeBridgeLink(BridgeBridgeLink link) {
-        BridgePort bp = new BridgePort();
-        bp.setNodeId(link.getNode().getId());
-        bp.setBridgePort(link.getBridgePort());
-        bp.setBridgePortIfIndex(link.getBridgePortIfIndex());
-        bp.setVlan(link.getVlan());
-        return bp;
-    }
-
-    public static BridgePort getFromDesignatedBridgeBridgeLink(BridgeBridgeLink link) {
-        BridgePort bp = new BridgePort();
-        bp.setNodeId(link.getDesignatedNode().getId());
-        bp.setBridgePort(link.getDesignatedPort());
-        bp.setBridgePortIfIndex(link.getDesignatedPortIfIndex());
-        bp.setVlan(link.getDesignatedVlan());
-        return bp;
-    }
-
-    public static List<BridgeBridgeLink> create(BridgePort designatedPort, Set<BridgePort> ports) {
-        OnmsNode designatedNode = new OnmsNode();
-        designatedNode.setId(designatedPort.getNodeId());
-        List<BridgeBridgeLink> links = new ArrayList<>();
-        for (BridgePort port:ports) {
-            if (port.equals(designatedPort)) {
-                continue;
-            }
-            BridgeBridgeLink link = new BridgeBridgeLink();
-            OnmsNode node = new OnmsNode();
-            node.setId(port.getNodeId());
-            link.setNode(node);
-            link.setBridgePort(port.getBridgePort());
-            link.setBridgePortIfIndex(port.getBridgePortIfIndex());
-            link.setVlan(port.getVlan());
-            link.setDesignatedNode(designatedNode);
-            link.setDesignatedPort(designatedPort.getBridgePort());
-            link.setDesignatedPortIfIndex(designatedPort.getBridgePortIfIndex());
-            link.setDesignatedVlan(designatedPort.getVlan());
-            links.add(link);
-        }
-        return links;
-
-    }
 
     public Date getCreateTime() {
         return m_createTime;
@@ -226,15 +64,11 @@ public class SharedSegment implements Topology{
         return m_designatedBridgeId;
     }
 
-    public BridgePort getDesignatedPort() throws BridgeTopologyException {
+    public BridgePort getDesignatedPort() {
         if (m_designatedBridgeId == null) {
-            throw new BridgeTopologyException("Designated Bridge NodeId cannot be null", this);
+            return null;
         }
-        BridgePort designatedbridge= getBridgePort(m_designatedBridgeId);
-        if (designatedbridge == null) {
-            throw new BridgeTopologyException("Designated BridgePort cannot be null", this);
-        }
-        return designatedbridge;
+       return getBridgePort(m_designatedBridgeId);
     }
 
     public boolean isEmpty() {
@@ -267,19 +101,10 @@ public class SharedSegment implements Topology{
         return m_macsOnSegment.contains(mac);
     }
 
-    public BridgePort getBridgePort(Integer nodeid) throws BridgeTopologyException {
+    public BridgePort getBridgePort(Integer nodeid) {
         if (nodeid == null)
             return null;
         for (BridgePort port: m_portsOnSegment) {
-            if (port == null) {
-                throw new BridgeTopologyException("Shared Segment: BridgePort cannot be null.", this);
-            }
-            if (port.getNodeId() == null) {
-                throw new BridgeTopologyException("Shared Segment: BridgePort nodeid cannot be null.", this);
-            }
-            if (port.getBridgePort() == null) {
-                throw new BridgeTopologyException("Shared Segment: BridgePort bridgeport cannot be null.", this);
-            }
             if ( port.getNodeId().intValue() == nodeid.intValue()) {
                 return port;
             }
