+++ features/enlinkd/service/impl/src/main/java/org/opennms/netmgt/enlinkd/service/impl/CdpTopologyServiceImpl.java
@@ -47,6 +47,7 @@ import org.opennms.netmgt.enlinkd.persistence.api.CdpLinkDao;
 import org.opennms.netmgt.enlinkd.service.api.CdpTopologyService;
 import org.opennms.netmgt.enlinkd.service.api.CompositeKey;
 import org.opennms.netmgt.enlinkd.service.api.TopologyConnection;
+import org.opennms.netmgt.enlinkd.service.api.TopologyService;
 import org.opennms.netmgt.model.OnmsNode;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -204,7 +205,7 @@ public class CdpTopologyServiceImpl extends TopologyServiceImpl implements CdpTo
 
             parsed.add(sourceLink.getId());
             parsed.add(targetLink.getId());
-            results.add(TopologyConnection.of(sourceLink, targetLink));
+            results.add(TopologyService.of(sourceLink, targetLink));
         }
         return results;
     }
