+++ features/enlinkd/service/impl/src/main/java/org/opennms/netmgt/enlinkd/service/impl/OspfTopologyServiceImpl.java
@@ -45,6 +45,7 @@ import org.opennms.netmgt.enlinkd.persistence.api.OspfLinkDao;
 import org.opennms.netmgt.enlinkd.service.api.CompositeKey;
 import org.opennms.netmgt.enlinkd.service.api.OspfTopologyService;
 import org.opennms.netmgt.enlinkd.service.api.TopologyConnection;
+import org.opennms.netmgt.enlinkd.service.api.TopologyService;
 import org.opennms.netmgt.model.OnmsNode;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -205,7 +206,7 @@ public class OspfTopologyServiceImpl extends TopologyServiceImpl implements Ospf
 
             LOG.debug("getOspfLinks: target: {}", targetLink);
             parsed.add(targetLink.getId());
-            results.add(TopologyConnection.of(sourceLink, targetLink));
+            results.add(TopologyService.of(sourceLink, targetLink));
         }
         return results;
