+++ features/enlinkd/tests/src/test/java/org/opennms/netmgt/enlinkd/EnLinkdTestHelper.java
@@ -62,6 +62,7 @@ import org.opennms.netmgt.enlinkd.service.api.BridgeTopologyException;
 import org.opennms.netmgt.enlinkd.service.api.BroadcastDomain;
 import org.opennms.netmgt.enlinkd.service.api.SharedSegment;
 import org.opennms.netmgt.enlinkd.service.api.BridgeForwardingTableEntry.BridgeDot1qTpFdbStatus;
+import org.opennms.netmgt.enlinkd.service.api.TopologyService;
 import org.opennms.netmgt.model.OnmsNode;
 import org.opennms.netmgt.model.monitoringLocations.OnmsMonitoringLocation;
 
@@ -1460,7 +1461,7 @@ public abstract class EnLinkdTestHelper {
                     fail();
                     break;
                 }
-                for (BridgeMacLink bridgeMacLink : SharedSegment.getBridgeMacLinks(segment)) {
+                for (BridgeMacLink bridgeMacLink : TopologyService.getBridgeMacLinks(segment)) {
                     assertNotNull(bridgeMacLink.getBridgePort());
                     assertNotNull(bridgeMacLink.getBridgePortIfIndex());
                     assertNotNull(bridgeMacLink.getMacAddress());
@@ -1498,9 +1499,9 @@ public abstract class EnLinkdTestHelper {
             switch (segment.getDesignatedPort().getBridgePort()) {
             case 1:
                 assertEquals(1, segment.getBridgePortsOnSegment().size());
-                assertEquals(0, SharedSegment.getBridgeBridgeLinks(segment).size());
+                assertEquals(0, TopologyService.getBridgeBridgeLinks(segment).size());
                 assertEquals(3, segment.getMacsOnSegment().size());
-                assertEquals(3, SharedSegment.getBridgeMacLinks(segment).size());
+                assertEquals(3, TopologyService.getBridgeMacLinks(segment).size());
                 break;
             case 2:
                 assertEquals(2, segment.getBridgePortsOnSegment().size());
@@ -1516,9 +1517,9 @@ public abstract class EnLinkdTestHelper {
                 for (String mac: macsonEport23) {
                     assertTrue(segment.containsMac(mac));
                 }
-                assertEquals(macsOnWirelessSegment.length+forwardersABCD.length+macsonEport23.length, SharedSegment.getBridgeMacLinks(segment).size());
-                assertEquals(1, SharedSegment.getBridgeBridgeLinks(segment).size());
-                BridgeBridgeLink blink = SharedSegment.getBridgeBridgeLinks(segment).iterator().next();  
+                assertEquals(macsOnWirelessSegment.length+forwardersABCD.length+macsonEport23.length, TopologyService.getBridgeMacLinks(segment).size());
+                assertEquals(1, TopologyService.getBridgeBridgeLinks(segment).size());
+                BridgeBridgeLink blink = TopologyService.getBridgeBridgeLinks(segment).iterator().next();
                 assertEquals(nodeBId, blink.getDesignatedNode().getId().intValue());
                 assertEquals(2, blink.getDesignatedPort().intValue());
                 assertEquals(11, blink.getBridgePort().intValue());
@@ -1584,7 +1585,7 @@ public abstract class EnLinkdTestHelper {
                     fail();
                     break;
                 }
-                for (BridgeMacLink bridgeMacLink : SharedSegment.getBridgeMacLinks(segment)) {
+                for (BridgeMacLink bridgeMacLink : TopologyService.getBridgeMacLinks(segment)) {
                     assertNotNull(bridgeMacLink.getBridgePort());
                     assertNotNull(bridgeMacLink.getBridgePortIfIndex());
                     assertNotNull(bridgeMacLink.getMacAddress());
@@ -1892,10 +1893,10 @@ public abstract class EnLinkdTestHelper {
             switch (segment.getDesignatedPort().getBridgePort()) {
             case 1:
                 assertEquals(2, segment.getBridgePortsOnSegment().size());
-                assertEquals(1, SharedSegment.getBridgeBridgeLinks(segment).size());
+                assertEquals(1, TopologyService.getBridgeBridgeLinks(segment).size());
                 assertEquals(0, segment.getMacsOnSegment().size());
-                assertEquals(0, SharedSegment.getBridgeMacLinks(segment).size());
-                BridgeBridgeLink link = SharedSegment.getBridgeBridgeLinks(segment).iterator().next();
+                assertEquals(0, TopologyService.getBridgeMacLinks(segment).size());
+                BridgeBridgeLink link = TopologyService.getBridgeBridgeLinks(segment).iterator().next();
                 assertEquals(nodeDId, link.getNode().getId().intValue());
                 assertEquals(50, link.getBridgePort().intValue());
                 assertEquals(nodeBId, link.getDesignatedNode().getId().intValue());
@@ -1910,9 +1911,9 @@ public abstract class EnLinkdTestHelper {
                 for (String mac: macsOnWirelessSegment) {
                     assertTrue(segment.containsMac(mac));
                 }
-                assertEquals(macsOnWirelessSegment.length, SharedSegment.getBridgeMacLinks(segment).size());
-                assertEquals(2, SharedSegment.getBridgeBridgeLinks(segment).size());
-                for (BridgeBridgeLink blink: SharedSegment.getBridgeBridgeLinks(segment))  {
+                assertEquals(macsOnWirelessSegment.length, TopologyService.getBridgeMacLinks(segment).size());
+                assertEquals(2, TopologyService.getBridgeBridgeLinks(segment).size());
+                for (BridgeBridgeLink blink: TopologyService.getBridgeBridgeLinks(segment))  {
                     assertEquals(nodeBId, blink.getDesignatedNode().getId().intValue());
                     assertEquals(2, blink.getDesignatedPort().intValue());
                     
@@ -1955,11 +1956,11 @@ public abstract class EnLinkdTestHelper {
         public void checkD(SharedSegment segment) throws BridgeTopologyException {
             assertEquals(nodeDId, segment.getDesignatedBridge().intValue());
             assertEquals(2, segment.getBridgePortsOnSegment().size());
-            assertEquals(1, SharedSegment.getBridgeBridgeLinks(segment).size());
+            assertEquals(1, TopologyService.getBridgeBridgeLinks(segment).size());
             assertEquals(0, segment.getMacsOnSegment().size());
-            assertEquals(0, SharedSegment.getBridgeMacLinks(segment).size());
+            assertEquals(0, TopologyService.getBridgeMacLinks(segment).size());
             assertEquals(49, segment.getDesignatedPort().getBridgePort().intValue());
-            BridgeBridgeLink link = SharedSegment.getBridgeBridgeLinks(segment).iterator().next();
+            BridgeBridgeLink link = TopologyService.getBridgeBridgeLinks(segment).iterator().next();
             assertEquals(nodeAId, link.getNode().getId().intValue());
             assertEquals(24, link.getBridgePort().intValue());
             assertEquals(nodeDId, link.getDesignatedNode().getId().intValue());
@@ -2063,8 +2064,8 @@ public abstract class EnLinkdTestHelper {
             assertEquals(3, shsegms.size());
 
             for (SharedSegment shared: shsegms) {
-                List<BridgeMacLink> links = SharedSegment.getBridgeMacLinks(shared);
-                List<BridgeBridgeLink> bblinks = SharedSegment.getBridgeBridgeLinks(shared);
+                List<BridgeMacLink> links = TopologyService.getBridgeMacLinks(shared);
+                List<BridgeBridgeLink> bblinks = TopologyService.getBridgeBridgeLinks(shared);
                 if (shared.getMacsOnSegment().contains(mac1)) {
                     assertEquals(0, bblinks.size());
                     assertEquals(1, shared.getBridgeIdsOnSegment().size());
@@ -2111,8 +2112,8 @@ public abstract class EnLinkdTestHelper {
             List<SharedSegment> shsegms = domain.getSharedSegments(); 
             assertEquals(4, shsegms.size());
             for (SharedSegment shared: shsegms) {
-                List<BridgeMacLink> links = SharedSegment.getBridgeMacLinks(shared);
-                List<BridgeBridgeLink> bblinks = SharedSegment.getBridgeBridgeLinks(shared);
+                List<BridgeMacLink> links = TopologyService.getBridgeMacLinks(shared);
+                List<BridgeBridgeLink> bblinks = TopologyService.getBridgeBridgeLinks(shared);
                 if (shared.noMacsOnSegment()) {
                     assertEquals(2, shared.getBridgeIdsOnSegment().size());
                     assertEquals(0, links.size());
@@ -2168,8 +2169,8 @@ public abstract class EnLinkdTestHelper {
             List<SharedSegment> shsegms = domain.getSharedSegments();
             assertEquals(4, shsegms.size());
             for (SharedSegment shared: shsegms) {
-                List<BridgeMacLink> links = SharedSegment.getBridgeMacLinks(shared);
-                List<BridgeBridgeLink> bblinks = SharedSegment.getBridgeBridgeLinks(shared);
+                List<BridgeMacLink> links = TopologyService.getBridgeMacLinks(shared);
+                List<BridgeBridgeLink> bblinks = TopologyService.getBridgeBridgeLinks(shared);
                 if (shared.noMacsOnSegment()) {
                     assertEquals(2, shared.getBridgeIdsOnSegment().size());
                     assertEquals(0, links.size());
@@ -2225,8 +2226,8 @@ public abstract class EnLinkdTestHelper {
             List<SharedSegment> shsegms = domain.getSharedSegments();
             assertEquals(5, shsegms.size());
             for (SharedSegment shared: shsegms) {
-                List<BridgeMacLink> links = SharedSegment.getBridgeMacLinks(shared);
-                List<BridgeBridgeLink> bblinks = SharedSegment.getBridgeBridgeLinks(shared);
+                List<BridgeMacLink> links = TopologyService.getBridgeMacLinks(shared);
+                List<BridgeBridgeLink> bblinks = TopologyService.getBridgeBridgeLinks(shared);
                 if (shared.noMacsOnSegment()) {
                     assertEquals(2, shared.getBridgeIdsOnSegment().size());
                     assertEquals(0, links.size());
@@ -2296,8 +2297,8 @@ public abstract class EnLinkdTestHelper {
             List<SharedSegment> shsegms = domain.getSharedSegments();
             assertEquals(5, shsegms.size());
             for (SharedSegment shared: shsegms) {
-                List<BridgeMacLink> links = SharedSegment.getBridgeMacLinks(shared);
-                List<BridgeBridgeLink> bblinks = SharedSegment.getBridgeBridgeLinks(shared);
+                List<BridgeMacLink> links = TopologyService.getBridgeMacLinks(shared);
+                List<BridgeBridgeLink> bblinks = TopologyService.getBridgeBridgeLinks(shared);
                 if (shared.noMacsOnSegment()) {
                     assertEquals(2, shared.getBridgeIdsOnSegment().size());
                     assertEquals(0, links.size());
@@ -2573,11 +2574,11 @@ public abstract class EnLinkdTestHelper {
                     assertEquals(1, shared.getBridgeIdsOnSegment().size());
                     assertEquals(2, shared.getMacsOnSegment().size());
                     assertTrue(shared.getMacsOnSegment().contains(mac2));
-                    assertEquals(2, SharedSegment.getBridgeMacLinks(shared).size());
-                    assertEquals(0, SharedSegment.getBridgeBridgeLinks(shared).size());
+                    assertEquals(2, TopologyService.getBridgeMacLinks(shared).size());
+                    assertEquals(0, TopologyService.getBridgeBridgeLinks(shared).size());
                     assertEquals(nodeDId.intValue(), shared.getDesignatedBridge().intValue());
                     assertEquals(portD.intValue(), shared.getDesignatedPort().getBridgePort().intValue());
-                    for (BridgeMacLink link: SharedSegment.getBridgeMacLinks(shared)) {
+                    for (BridgeMacLink link: TopologyService.getBridgeMacLinks(shared)) {
                        assertEquals(nodeDId.intValue(), link.getNode().getId().intValue());
                        assertEquals(portD,link.getBridgePort());
                        assertEquals(BridgeMacLinkType.BRIDGE_LINK, link.getLinkType());
@@ -2586,11 +2587,11 @@ public abstract class EnLinkdTestHelper {
                     assertEquals(1, shared.getBridgeIdsOnSegment().size());
                     assertEquals(2, shared.getMacsOnSegment().size());
                     assertTrue(shared.getMacsOnSegment().contains(mac6));
-                    assertEquals(2, SharedSegment.getBridgeMacLinks(shared).size());
-                    assertEquals(0, SharedSegment.getBridgeBridgeLinks(shared).size());
+                    assertEquals(2, TopologyService.getBridgeMacLinks(shared).size());
+                    assertEquals(0, TopologyService.getBridgeBridgeLinks(shared).size());
                     assertEquals(nodeEId.intValue(), shared.getDesignatedBridge().intValue());
                     assertEquals(portE.intValue(), shared.getDesignatedPort().getBridgePort().intValue());
-                    for (BridgeMacLink link: SharedSegment.getBridgeMacLinks(shared)) {
+                    for (BridgeMacLink link: TopologyService.getBridgeMacLinks(shared)) {
                         assertEquals(nodeEId.intValue(), link.getNode().getId().intValue());
                         assertEquals(portE,link.getBridgePort());
                         assertEquals(BridgeMacLinkType.BRIDGE_LINK, link.getLinkType());
@@ -2603,15 +2604,15 @@ public abstract class EnLinkdTestHelper {
                     assertTrue(shared.getMacsOnSegment().contains(mac3));
                     assertTrue(shared.getMacsOnSegment().contains(mac4));
                     assertTrue(shared.getMacsOnSegment().contains(mac8));
-                    assertEquals(4, SharedSegment.getBridgeMacLinks(shared).size());
-                    assertEquals(1, SharedSegment.getBridgeBridgeLinks(shared).size());
-                    BridgeBridgeLink bblink = SharedSegment.getBridgeBridgeLinks(shared).iterator().next();
+                    assertEquals(4, TopologyService.getBridgeMacLinks(shared).size());
+                    assertEquals(1, TopologyService.getBridgeBridgeLinks(shared).size());
+                    BridgeBridgeLink bblink = TopologyService.getBridgeBridgeLinks(shared).iterator().next();
                     assertEquals(nodeDId.intValue(), bblink.getDesignatedNode().getId().intValue());
                     assertEquals(portDD.intValue(), bblink.getDesignatedPort().intValue());
                     assertEquals(nodeEId.intValue(), bblink.getNode().getId().intValue());
                     assertEquals(portEE.intValue(), bblink.getBridgePort().intValue());
                     Set<String> macs = new HashSet<>();
-                    for (BridgeMacLink link: SharedSegment.getBridgeMacLinks(shared)) {
+                    for (BridgeMacLink link: TopologyService.getBridgeMacLinks(shared)) {
                         assertEquals(nodeDId.intValue(),link.getNode().getId().intValue());
                         assertEquals(portDD,link.getBridgePort());
                         assertFalse(macs.contains(link.getMacAddress()));
@@ -2639,11 +2640,11 @@ public abstract class EnLinkdTestHelper {
                     assertEquals(1, shared.getBridgeIdsOnSegment().size());
                     assertEquals(2, shared.getMacsOnSegment().size());
                     assertTrue(shared.getMacsOnSegment().contains(mac2));
-                    assertEquals(2, SharedSegment.getBridgeMacLinks(shared).size());
-                    assertEquals(0, SharedSegment.getBridgeBridgeLinks(shared).size());
+                    assertEquals(2, TopologyService.getBridgeMacLinks(shared).size());
+                    assertEquals(0, TopologyService.getBridgeBridgeLinks(shared).size());
                     assertEquals(nodeDId.intValue(), shared.getDesignatedBridge().intValue());
                     assertEquals(portD.intValue(), shared.getDesignatedPort().getBridgePort().intValue());
-                    for (BridgeMacLink link: SharedSegment.getBridgeMacLinks(shared)) {
+                    for (BridgeMacLink link: TopologyService.getBridgeMacLinks(shared)) {
                        assertEquals(nodeDId.intValue(), link.getNode().getId().intValue());
                        assertEquals(portD,link.getBridgePort());
                        assertEquals(BridgeMacLinkType.BRIDGE_LINK, link.getLinkType());
@@ -2652,11 +2653,11 @@ public abstract class EnLinkdTestHelper {
                     assertEquals(1, shared.getBridgeIdsOnSegment().size());
                     assertEquals(2, shared.getMacsOnSegment().size());
                     assertTrue(shared.getMacsOnSegment().contains(mac4));
-                    assertEquals(2, SharedSegment.getBridgeMacLinks(shared).size());
-                    assertEquals(0, SharedSegment.getBridgeBridgeLinks(shared).size());
+                    assertEquals(2, TopologyService.getBridgeMacLinks(shared).size());
+                    assertEquals(0, TopologyService.getBridgeBridgeLinks(shared).size());
                     assertEquals(nodeFId.intValue(), shared.getDesignatedBridge().intValue());
                     assertEquals(portF.intValue(), shared.getDesignatedPort().getBridgePort().intValue());
-                    for (BridgeMacLink link: SharedSegment.getBridgeMacLinks(shared)) {
+                    for (BridgeMacLink link: TopologyService.getBridgeMacLinks(shared)) {
                         assertEquals(nodeFId.intValue(), link.getNode().getId().intValue());
                         assertEquals(portF,link.getBridgePort());
                         assertEquals(BridgeMacLinkType.BRIDGE_LINK, link.getLinkType());
@@ -2669,15 +2670,15 @@ public abstract class EnLinkdTestHelper {
                     assertTrue(shared.getMacsOnSegment().contains(mac5));
                     assertTrue(shared.getMacsOnSegment().contains(mac6));
                     assertTrue(shared.getMacsOnSegment().contains(mac8));
-                    assertEquals(4, SharedSegment.getBridgeMacLinks(shared).size());
-                    BridgeBridgeLink bblink = SharedSegment.getBridgeBridgeLinks(shared).iterator().next();
+                    assertEquals(4, TopologyService.getBridgeMacLinks(shared).size());
+                    BridgeBridgeLink bblink = TopologyService.getBridgeBridgeLinks(shared).iterator().next();
                     assertEquals(nodeDId.intValue(), bblink.getDesignatedNode().getId().intValue());
                     assertEquals(portDD.intValue(), bblink.getDesignatedPort().intValue());
                     assertEquals(nodeFId.intValue(), bblink.getNode().getId().intValue());
                     assertEquals(portFF.intValue(), bblink.getBridgePort().intValue());
-                    assertEquals(1, SharedSegment.getBridgeBridgeLinks(shared).size());
+                    assertEquals(1, TopologyService.getBridgeBridgeLinks(shared).size());
                     Set<String> macs = new HashSet<>();
-                    for (BridgeMacLink link: SharedSegment.getBridgeMacLinks(shared)) {
+                    for (BridgeMacLink link: TopologyService.getBridgeMacLinks(shared)) {
                         assertEquals(nodeDId.intValue(),link.getNode().getId().intValue());
                         assertEquals(portDD,link.getBridgePort());
                         assertFalse(macs.contains(link.getMacAddress()));
@@ -2705,9 +2706,9 @@ public abstract class EnLinkdTestHelper {
                     assertEquals(2, shared.getBridgeIdsOnSegment().size());
                     assertTrue(shared.getBridgeIdsOnSegment().contains(nodeGId));
                     assertTrue(shared.getBridgeIdsOnSegment().contains(nodeDId));
-                    assertEquals(1, SharedSegment.getBridgeBridgeLinks(shared).size());
-                    assertEquals(0, SharedSegment.getBridgeMacLinks(shared).size());
-                    BridgeBridgeLink link = SharedSegment.getBridgeBridgeLinks(shared).iterator().next();
+                    assertEquals(1, TopologyService.getBridgeBridgeLinks(shared).size());
+                    assertEquals(0, TopologyService.getBridgeMacLinks(shared).size());
+                    BridgeBridgeLink link = TopologyService.getBridgeBridgeLinks(shared).iterator().next();
                     assertEquals(nodeDId, link.getDesignatedNode().getId());
                     assertEquals(nodeGId, link.getNode().getId());
                     assertEquals(portDD, link.getDesignatedPort());
@@ -2716,11 +2717,11 @@ public abstract class EnLinkdTestHelper {
                     assertEquals(1, shared.getBridgeIdsOnSegment().size());
                     assertEquals(2, shared.getMacsOnSegment().size());
                     assertTrue(shared.getMacsOnSegment().contains(mac2));
-                    assertEquals(2, SharedSegment.getBridgeMacLinks(shared).size());
-                    assertEquals(0, SharedSegment.getBridgeBridgeLinks(shared).size());
+                    assertEquals(2, TopologyService.getBridgeMacLinks(shared).size());
+                    assertEquals(0, TopologyService.getBridgeBridgeLinks(shared).size());
                     assertEquals(nodeDId.intValue(), shared.getDesignatedBridge().intValue());
                     assertEquals(portD.intValue(), shared.getDesignatedPort().getBridgePort().intValue());
-                    for (BridgeMacLink link: SharedSegment.getBridgeMacLinks(shared)) {
+                    for (BridgeMacLink link: TopologyService.getBridgeMacLinks(shared)) {
                        assertEquals(nodeDId.intValue(), link.getNode().getId().intValue());
                        assertEquals(portD,link.getBridgePort());
                        assertEquals(BridgeMacLinkType.BRIDGE_LINK, link.getLinkType());
@@ -2729,11 +2730,11 @@ public abstract class EnLinkdTestHelper {
                     assertEquals(1, shared.getBridgeIdsOnSegment().size());
                     assertEquals(2, shared.getMacsOnSegment().size());
                     assertTrue(shared.getMacsOnSegment().contains(mac4));
-                    assertEquals(2, SharedSegment.getBridgeMacLinks(shared).size());
-                    assertEquals(0, SharedSegment.getBridgeBridgeLinks(shared).size());
+                    assertEquals(2, TopologyService.getBridgeMacLinks(shared).size());
+                    assertEquals(0, TopologyService.getBridgeBridgeLinks(shared).size());
                     assertEquals(nodeGId.intValue(), shared.getDesignatedBridge().intValue());
                     assertEquals(portGF.intValue(), shared.getDesignatedPort().getBridgePort().intValue());
-                    for (BridgeMacLink link: SharedSegment.getBridgeMacLinks(shared)) {
+                    for (BridgeMacLink link: TopologyService.getBridgeMacLinks(shared)) {
                         assertEquals(nodeGId.intValue(), link.getNode().getId().intValue());
                         assertEquals(portGF,link.getBridgePort());
                         assertEquals(BridgeMacLinkType.BRIDGE_LINK, link.getLinkType());
@@ -2742,11 +2743,11 @@ public abstract class EnLinkdTestHelper {
                     assertEquals(1, shared.getBridgeIdsOnSegment().size());
                     assertEquals(2, shared.getMacsOnSegment().size());
                     assertTrue(shared.getMacsOnSegment().contains(mac6));
-                    assertEquals(2, SharedSegment.getBridgeMacLinks(shared).size());
-                    assertEquals(0, SharedSegment.getBridgeBridgeLinks(shared).size());
+                    assertEquals(2, TopologyService.getBridgeMacLinks(shared).size());
+                    assertEquals(0, TopologyService.getBridgeBridgeLinks(shared).size());
                     assertEquals(nodeGId.intValue(), shared.getDesignatedBridge().intValue());
                     assertEquals(portGE.intValue(), shared.getDesignatedPort().getBridgePort().intValue());
-                    for (BridgeMacLink link: SharedSegment.getBridgeMacLinks(shared)) {
+                    for (BridgeMacLink link: TopologyService.getBridgeMacLinks(shared)) {
                         assertEquals(nodeGId.intValue(), link.getNode().getId().intValue());
                         assertEquals(portGE,link.getBridgePort());
                         assertEquals(BridgeMacLinkType.BRIDGE_LINK, link.getLinkType());
@@ -2757,7 +2758,7 @@ public abstract class EnLinkdTestHelper {
                     assertEquals(1,shared.getMacsOnSegment().size());
                     assertEquals(nodeGId.intValue(), shared.getDesignatedBridge().intValue());
                     assertEquals(portG7.intValue(), shared.getDesignatedPort().getBridgePort().intValue());
-                    for (BridgeMacLink link: SharedSegment.getBridgeMacLinks(shared)) {
+                    for (BridgeMacLink link: TopologyService.getBridgeMacLinks(shared)) {
                             assertEquals(portG7,link.getBridgePort());
                             assertEquals(nodeGId, link.getNode().getId());
                             assertEquals(BridgeMacLinkType.BRIDGE_LINK, link.getLinkType());
@@ -2768,7 +2769,7 @@ public abstract class EnLinkdTestHelper {
                      assertEquals(1,shared.getMacsOnSegment().size());
                      assertEquals(nodeGId.intValue(), shared.getDesignatedBridge().intValue());
                      assertEquals(portG8.intValue(), shared.getDesignatedPort().getBridgePort().intValue());
-                     for (BridgeMacLink link: SharedSegment.getBridgeMacLinks(shared)) {
+                     for (BridgeMacLink link: TopologyService.getBridgeMacLinks(shared)) {
                              assertEquals(portG8,link.getBridgePort());
                              assertEquals(nodeGId, link.getNode().getId());
                              assertEquals(BridgeMacLinkType.BRIDGE_LINK, link.getLinkType());
@@ -2790,12 +2791,12 @@ public abstract class EnLinkdTestHelper {
                     assertEquals(1, shared.getBridgeIdsOnSegment().size());
                     assertEquals(2, shared.getMacsOnSegment().size());
                     assertTrue(shared.getMacsOnSegment().contains(mac4));
-                    assertEquals(2, SharedSegment.getBridgeMacLinks(shared).size());
-                    assertEquals(0, SharedSegment.getBridgeBridgeLinks(shared).size());
+                    assertEquals(2, TopologyService.getBridgeMacLinks(shared).size());
+                    assertEquals(0, TopologyService.getBridgeBridgeLinks(shared).size());
                     assertEquals(nodeFId.intValue(), shared.getDesignatedBridge().intValue());
                     assertEquals(portF.intValue(), shared.getDesignatedPort().getBridgePort().intValue());
                     Set<String> macs = new HashSet<>();
-                    for (BridgeMacLink link: SharedSegment.getBridgeMacLinks(shared)) {
+                    for (BridgeMacLink link: TopologyService.getBridgeMacLinks(shared)) {
                         assertEquals(nodeFId.intValue(), link.getNode().getId().intValue());
                         assertEquals(portF,link.getBridgePort());
                         assertFalse(macs.contains(link.getMacAddress()));
@@ -2808,12 +2809,12 @@ public abstract class EnLinkdTestHelper {
                     assertEquals(1, shared.getBridgeIdsOnSegment().size());
                     assertEquals(2, shared.getMacsOnSegment().size());
                     assertTrue(shared.getMacsOnSegment().contains(mac6));
-                    assertEquals(2, SharedSegment.getBridgeMacLinks(shared).size());
-                    assertEquals(0, SharedSegment.getBridgeBridgeLinks(shared).size());
+                    assertEquals(2, TopologyService.getBridgeMacLinks(shared).size());
+                    assertEquals(0, TopologyService.getBridgeBridgeLinks(shared).size());
                     assertEquals(nodeEId.intValue(), shared.getDesignatedBridge().intValue());
                     assertEquals(portE.intValue(), shared.getDesignatedPort().getBridgePort().intValue());
                     Set<String> macs = new HashSet<>();
-                    for (BridgeMacLink link: SharedSegment.getBridgeMacLinks(shared)) {
+                    for (BridgeMacLink link: TopologyService.getBridgeMacLinks(shared)) {
                         assertEquals(nodeEId.intValue(), link.getNode().getId().intValue());
                         assertEquals(portE,link.getBridgePort());
                         assertFalse(macs.contains(link.getMacAddress()));
@@ -2830,15 +2831,15 @@ public abstract class EnLinkdTestHelper {
                     assertTrue(shared.getMacsOnSegment().contains(mac2));
                     assertTrue(shared.getMacsOnSegment().contains(mac7));
                     assertTrue(shared.getMacsOnSegment().contains(mac8));
-                    assertEquals(1, SharedSegment.getBridgeBridgeLinks(shared).size());
-                    BridgeBridgeLink bblink = SharedSegment.getBridgeBridgeLinks(shared).iterator().next();
+                    assertEquals(1, TopologyService.getBridgeBridgeLinks(shared).size());
+                    BridgeBridgeLink bblink = TopologyService.getBridgeBridgeLinks(shared).iterator().next();
                     assertEquals(nodeEId.intValue(), bblink.getDesignatedNode().getId().intValue());
                     assertEquals(portEE, bblink.getDesignatedPort());
                     assertEquals(nodeFId.intValue(), bblink.getNode().getId().intValue());
                     assertEquals(portFF, bblink.getBridgePort());
-                    assertEquals(4, SharedSegment.getBridgeMacLinks(shared).size());
+                    assertEquals(4, TopologyService.getBridgeMacLinks(shared).size());
                     Set<String> macs = new HashSet<>();
-                    for (BridgeMacLink link: SharedSegment.getBridgeMacLinks(shared)) {
+                    for (BridgeMacLink link: TopologyService.getBridgeMacLinks(shared)) {
                         assertEquals(nodeEId.intValue(), link.getNode().getId().intValue());
                         assertEquals(portEE,link.getBridgePort());
                         assertFalse(macs.contains(link.getMacAddress()));
@@ -2868,12 +2869,12 @@ public abstract class EnLinkdTestHelper {
                     assertEquals(1, shared.getBridgeIdsOnSegment().size());
                     assertEquals(2, shared.getMacsOnSegment().size());
                     assertTrue(shared.getMacsOnSegment().contains(mac2));
-                    assertEquals(2, SharedSegment.getBridgeMacLinks(shared).size());
-                    assertEquals(0, SharedSegment.getBridgeBridgeLinks(shared).size());
+                    assertEquals(2, TopologyService.getBridgeMacLinks(shared).size());
+                    assertEquals(0, TopologyService.getBridgeBridgeLinks(shared).size());
                     assertEquals(nodeDId.intValue(), shared.getDesignatedBridge().intValue());
                     assertEquals(portD.intValue(), shared.getDesignatedPort().getBridgePort().intValue());
                     Set<String> macs = new HashSet<>();
-                    for (BridgeMacLink link: SharedSegment.getBridgeMacLinks(shared)) {
+                    for (BridgeMacLink link: TopologyService.getBridgeMacLinks(shared)) {
                        assertEquals(nodeDId.intValue(), link.getNode().getId().intValue());
                        assertEquals(portD,link.getBridgePort());
                        assertFalse(macs.contains(link.getMacAddress()));
@@ -2886,12 +2887,12 @@ public abstract class EnLinkdTestHelper {
                     assertEquals(1, shared.getBridgeIdsOnSegment().size());
                     assertEquals(2, shared.getMacsOnSegment().size());
                     assertTrue(shared.getMacsOnSegment().contains(mac6));
-                    assertEquals(2, SharedSegment.getBridgeMacLinks(shared).size());
-                    assertEquals(0, SharedSegment.getBridgeBridgeLinks(shared).size());
+                    assertEquals(2, TopologyService.getBridgeMacLinks(shared).size());
+                    assertEquals(0, TopologyService.getBridgeBridgeLinks(shared).size());
                     assertEquals(nodeEId.intValue(), shared.getDesignatedBridge().intValue());
                     assertEquals(portE.intValue(), shared.getDesignatedPort().getBridgePort().intValue());
                     Set<String> macs = new HashSet<>();
-                    for (BridgeMacLink link: SharedSegment.getBridgeMacLinks(shared)) {
+                    for (BridgeMacLink link: TopologyService.getBridgeMacLinks(shared)) {
                         assertEquals(nodeEId.intValue(), link.getNode().getId().intValue());
                         assertEquals(portE,link.getBridgePort());
                         assertFalse(macs.contains(link.getMacAddress()));
@@ -2904,12 +2905,12 @@ public abstract class EnLinkdTestHelper {
                     assertEquals(1, shared.getBridgeIdsOnSegment().size());
                     assertEquals(2, shared.getMacsOnSegment().size());
                     assertTrue(shared.getMacsOnSegment().contains(mac4));
-                    assertEquals(2, SharedSegment.getBridgeMacLinks(shared).size());
-                    assertEquals(0, SharedSegment.getBridgeBridgeLinks(shared).size());
+                    assertEquals(2, TopologyService.getBridgeMacLinks(shared).size());
+                    assertEquals(0, TopologyService.getBridgeBridgeLinks(shared).size());
                     assertEquals(nodeFId.intValue(), shared.getDesignatedBridge().intValue());
                     assertEquals(portF.intValue(), shared.getDesignatedPort().getBridgePort().intValue());
                     Set<String> macs = new HashSet<>();
-                    for (BridgeMacLink link: SharedSegment.getBridgeMacLinks(shared)) {
+                    for (BridgeMacLink link: TopologyService.getBridgeMacLinks(shared)) {
                         assertEquals(nodeFId.intValue(), link.getNode().getId().intValue());
                         assertEquals(portF,link.getBridgePort());
                         assertFalse(macs.contains(link.getMacAddress()));
@@ -2925,10 +2926,10 @@ public abstract class EnLinkdTestHelper {
                     assertTrue(shared.getBridgeIdsOnSegment().contains(nodeFId));
                     assertEquals(2,shared.getMacsOnSegment().size());
                     assertTrue(shared.getMacsOnSegment().contains(mac8));
-                    assertEquals(2, SharedSegment.getBridgeMacLinks(shared).size());
-                    assertEquals(2, SharedSegment.getBridgeBridgeLinks(shared).size());
+                    assertEquals(2, TopologyService.getBridgeMacLinks(shared).size());
+                    assertEquals(2, TopologyService.getBridgeBridgeLinks(shared).size());
                     Set<String> macs = new HashSet<>();
-                    for (BridgeMacLink link: SharedSegment.getBridgeMacLinks(shared)) {
+                    for (BridgeMacLink link: TopologyService.getBridgeMacLinks(shared)) {
                         assertEquals(nodeDId.intValue(), link.getNode().getId().intValue());
                         assertEquals(portDD,link.getBridgePort());
                         assertFalse(macs.contains(link.getMacAddress()));
@@ -2954,9 +2955,9 @@ public abstract class EnLinkdTestHelper {
                 if (shared.noMacsOnSegment()) {
                     assertEquals(2, shared.getBridgeIdsOnSegment().size());
                     assertTrue(shared.getBridgeIdsOnSegment().contains(nodeGId));
-                    assertEquals(1, SharedSegment.getBridgeBridgeLinks(shared).size());
-                    assertEquals(0, SharedSegment.getBridgeMacLinks(shared).size());
-                    BridgeBridgeLink link = SharedSegment.getBridgeBridgeLinks(shared).iterator().next();
+                    assertEquals(1, TopologyService.getBridgeBridgeLinks(shared).size());
+                    assertEquals(0, TopologyService.getBridgeMacLinks(shared).size());
+                    BridgeBridgeLink link = TopologyService.getBridgeBridgeLinks(shared).iterator().next();
                     if (shared.getBridgeIdsOnSegment().contains(nodeDId)) {
                         assertEquals(nodeDId.intValue(),link.getDesignatedNode().getId().intValue());
                         assertEquals(portDD,link.getDesignatedPort());
@@ -2980,12 +2981,12 @@ public abstract class EnLinkdTestHelper {
                     assertEquals(1, shared.getBridgeIdsOnSegment().size());
                     assertEquals(2, shared.getMacsOnSegment().size());
                     assertTrue(shared.getMacsOnSegment().contains(mac2));
-                    assertEquals(2, SharedSegment.getBridgeMacLinks(shared).size());
-                    assertEquals(0, SharedSegment.getBridgeBridgeLinks(shared).size());
+                    assertEquals(2, TopologyService.getBridgeMacLinks(shared).size());
+                    assertEquals(0, TopologyService.getBridgeBridgeLinks(shared).size());
                     assertEquals(nodeDId.intValue(), shared.getDesignatedBridge().intValue());
                     assertEquals(portD.intValue(), shared.getDesignatedPort().getBridgePort().intValue());
                     Set<String> macs = new HashSet<>();
-                    for (BridgeMacLink link: SharedSegment.getBridgeMacLinks(shared)) {
+                    for (BridgeMacLink link: TopologyService.getBridgeMacLinks(shared)) {
                        assertEquals(nodeDId.intValue(), link.getNode().getId().intValue());
                        assertEquals(portD,link.getBridgePort());
                        assertFalse(macs.contains(link.getMacAddress()));
@@ -2999,12 +3000,12 @@ public abstract class EnLinkdTestHelper {
                     assertEquals(1, shared.getBridgeIdsOnSegment().size());
                     assertEquals(2, shared.getMacsOnSegment().size());
                     assertTrue(shared.getMacsOnSegment().contains(mac6));
-                    assertEquals(2, SharedSegment.getBridgeMacLinks(shared).size());
-                    assertEquals(0, SharedSegment.getBridgeBridgeLinks(shared).size());
+                    assertEquals(2, TopologyService.getBridgeMacLinks(shared).size());
+                    assertEquals(0, TopologyService.getBridgeBridgeLinks(shared).size());
                     assertEquals(nodeEId.intValue(), shared.getDesignatedBridge().intValue());
                     assertEquals(portE.intValue(), shared.getDesignatedPort().getBridgePort().intValue());
                     Set<String> macs = new HashSet<>();
-                    for (BridgeMacLink link: SharedSegment.getBridgeMacLinks(shared)) {
+                    for (BridgeMacLink link: TopologyService.getBridgeMacLinks(shared)) {
                         assertEquals(nodeEId.intValue(), link.getNode().getId().intValue());
                         assertEquals(portE,link.getBridgePort());
                         assertFalse(macs.contains(link.getMacAddress()));
@@ -3017,12 +3018,12 @@ public abstract class EnLinkdTestHelper {
                     assertEquals(1, shared.getBridgeIdsOnSegment().size());
                     assertEquals(2, shared.getMacsOnSegment().size());
                     assertTrue(shared.getMacsOnSegment().contains(mac4));
-                    assertEquals(2, SharedSegment.getBridgeMacLinks(shared).size());
-                    assertEquals(0, SharedSegment.getBridgeBridgeLinks(shared).size());
+                    assertEquals(2, TopologyService.getBridgeMacLinks(shared).size());
+                    assertEquals(0, TopologyService.getBridgeBridgeLinks(shared).size());
                     assertEquals(nodeFId.intValue(), shared.getDesignatedBridge().intValue());
                     assertEquals(portF.intValue(), shared.getDesignatedPort().getBridgePort().intValue());
                     Set<String> macs = new HashSet<>();
-                    for (BridgeMacLink link: SharedSegment.getBridgeMacLinks(shared)) {
+                    for (BridgeMacLink link: TopologyService.getBridgeMacLinks(shared)) {
                         assertEquals(nodeFId.intValue(), link.getNode().getId().intValue());
                         assertEquals(portF,link.getBridgePort());
                         assertFalse(macs.contains(link.getMacAddress()));
@@ -3035,9 +3036,9 @@ public abstract class EnLinkdTestHelper {
                     assertEquals(1, shared.getBridgeIdsOnSegment().size());
                     assertTrue(shared.getBridgeIdsOnSegment().contains(nodeGId));
                     assertEquals(1,shared.getMacsOnSegment().size());
-                    assertEquals(1, SharedSegment.getBridgeMacLinks(shared).size());
-                    assertEquals(0, SharedSegment.getBridgeBridgeLinks(shared).size());
-                    BridgeMacLink link = SharedSegment.getBridgeMacLinks(shared).iterator().next();
+                    assertEquals(1, TopologyService.getBridgeMacLinks(shared).size());
+                    assertEquals(0, TopologyService.getBridgeBridgeLinks(shared).size());
+                    BridgeMacLink link = TopologyService.getBridgeMacLinks(shared).iterator().next();
                     assertEquals(nodeGId.intValue(), link.getNode().getId().intValue());
                     assertEquals(portG7.intValue(), link.getBridgePort().intValue());
                     assertEquals(mac7, link.getMacAddress());
@@ -3046,9 +3047,9 @@ public abstract class EnLinkdTestHelper {
                      assertEquals(1, shared.getBridgeIdsOnSegment().size());
                      assertTrue(shared.getBridgeIdsOnSegment().contains(nodeGId));
                      assertEquals(1,shared.getMacsOnSegment().size());
-                     assertEquals(1, SharedSegment.getBridgeMacLinks(shared).size());
-                     assertEquals(0, SharedSegment.getBridgeBridgeLinks(shared).size());
-                     BridgeMacLink link = SharedSegment.getBridgeMacLinks(shared).iterator().next();
+                     assertEquals(1, TopologyService.getBridgeMacLinks(shared).size());
+                     assertEquals(0, TopologyService.getBridgeBridgeLinks(shared).size());
+                     BridgeMacLink link = TopologyService.getBridgeMacLinks(shared).iterator().next();
                      assertEquals(nodeGId.intValue(), link.getNode().getId().intValue());
                      assertEquals(portG8.intValue(), link.getBridgePort().intValue());
                      assertEquals(mac8, link.getMacAddress());
@@ -3066,9 +3067,9 @@ public abstract class EnLinkdTestHelper {
             for (SharedSegment shared: shsegs) {
                 if (shared.noMacsOnSegment()) {
                     assertEquals(2, shared.getBridgeIdsOnSegment().size());
-                    assertEquals(1, SharedSegment.getBridgeBridgeLinks(shared).size());
-                    assertEquals(0, SharedSegment.getBridgeMacLinks(shared).size());
-                    BridgeBridgeLink link = SharedSegment.getBridgeBridgeLinks(shared).iterator().next();
+                    assertEquals(1, TopologyService.getBridgeBridgeLinks(shared).size());
+                    assertEquals(0, TopologyService.getBridgeMacLinks(shared).size());
+                    BridgeBridgeLink link = TopologyService.getBridgeBridgeLinks(shared).iterator().next();
                     if (shared.getBridgeIdsOnSegment().contains(nodeDId) && shared.getDesignatedPort().getBridgePort().intValue() == portDD.intValue()) {
                         assertEquals(nodeDId.intValue(),link.getDesignatedNode().getId().intValue());
                         assertEquals(portDD,link.getDesignatedPort());
@@ -3107,11 +3108,11 @@ public abstract class EnLinkdTestHelper {
                     assertEquals(1, shared.getBridgeIdsOnSegment().size());
                     assertTrue(shared.getBridgeIdsOnSegment().contains(nodeHId));
                     assertEquals(1, shared.getMacsOnSegment().size());
-                    assertEquals(1, SharedSegment.getBridgeMacLinks(shared).size());
-                    assertEquals(0, SharedSegment.getBridgeBridgeLinks(shared).size());
+                    assertEquals(1, TopologyService.getBridgeMacLinks(shared).size());
+                    assertEquals(0, TopologyService.getBridgeBridgeLinks(shared).size());
                     assertEquals(nodeHId.intValue(), shared.getDesignatedBridge().intValue());
                     assertEquals(portH1.intValue(), shared.getDesignatedPort().getBridgePort().intValue());
-                    BridgeMacLink link = SharedSegment.getBridgeMacLinks(shared).iterator().next();
+                    BridgeMacLink link = TopologyService.getBridgeMacLinks(shared).iterator().next();
                     assertEquals(nodeHId.intValue(), link.getNode().getId().intValue());
                     assertEquals(portH1,link.getBridgePort());
                     assertEquals(mac1,link.getMacAddress());
@@ -3120,11 +3121,11 @@ public abstract class EnLinkdTestHelper {
                     assertEquals(1, shared.getBridgeIdsOnSegment().size());
                     assertTrue(shared.getBridgeIdsOnSegment().contains(nodeHId));
                     assertEquals(1, shared.getMacsOnSegment().size());
-                    assertEquals(1, SharedSegment.getBridgeMacLinks(shared).size());
-                    assertEquals(0, SharedSegment.getBridgeBridgeLinks(shared).size());
+                    assertEquals(1, TopologyService.getBridgeMacLinks(shared).size());
+                    assertEquals(0, TopologyService.getBridgeBridgeLinks(shared).size());
                     assertEquals(nodeHId.intValue(), shared.getDesignatedBridge().intValue());
                     assertEquals(portH2.intValue(), shared.getDesignatedPort().getBridgePort().intValue());
-                    BridgeMacLink link = SharedSegment.getBridgeMacLinks(shared).iterator().next();
+                    BridgeMacLink link = TopologyService.getBridgeMacLinks(shared).iterator().next();
                     assertEquals(nodeHId.intValue(), link.getNode().getId().intValue());
                     assertEquals(portH2,link.getBridgePort());
                     assertEquals(mac2,link.getMacAddress());
@@ -3133,11 +3134,11 @@ public abstract class EnLinkdTestHelper {
                     assertEquals(1, shared.getBridgeIdsOnSegment().size());
                     assertTrue(shared.getBridgeIdsOnSegment().contains(nodeIId));
                     assertEquals(1, shared.getMacsOnSegment().size());
-                    assertEquals(1, SharedSegment.getBridgeMacLinks(shared).size());
-                    assertEquals(0, SharedSegment.getBridgeBridgeLinks(shared).size());
+                    assertEquals(1, TopologyService.getBridgeMacLinks(shared).size());
+                    assertEquals(0, TopologyService.getBridgeBridgeLinks(shared).size());
                     assertEquals(nodeIId.intValue(), shared.getDesignatedBridge().intValue());
                     assertEquals(portI3.intValue(), shared.getDesignatedPort().getBridgePort().intValue());
-                    BridgeMacLink link = SharedSegment.getBridgeMacLinks(shared).iterator().next();
+                    BridgeMacLink link = TopologyService.getBridgeMacLinks(shared).iterator().next();
                     assertEquals(nodeIId.intValue(), link.getNode().getId().intValue());
                     assertEquals(portI3,link.getBridgePort());
                     assertEquals(mac3,link.getMacAddress());
@@ -3146,11 +3147,11 @@ public abstract class EnLinkdTestHelper {
                     assertEquals(1, shared.getBridgeIdsOnSegment().size());
                     assertTrue(shared.getBridgeIdsOnSegment().contains(nodeIId));
                     assertEquals(1, shared.getMacsOnSegment().size());
-                    assertEquals(1, SharedSegment.getBridgeMacLinks(shared).size());
-                    assertEquals(0, SharedSegment.getBridgeBridgeLinks(shared).size());
+                    assertEquals(1, TopologyService.getBridgeMacLinks(shared).size());
+                    assertEquals(0, TopologyService.getBridgeBridgeLinks(shared).size());
                     assertEquals(nodeIId.intValue(), shared.getDesignatedBridge().intValue());
                     assertEquals(portI4.intValue(), shared.getDesignatedPort().getBridgePort().intValue());
-                    BridgeMacLink link = SharedSegment.getBridgeMacLinks(shared).iterator().next();
+                    BridgeMacLink link = TopologyService.getBridgeMacLinks(shared).iterator().next();
                     assertEquals(nodeIId.intValue(), link.getNode().getId().intValue());
                     assertEquals(portI4,link.getBridgePort());
                     assertEquals(mac4,link.getMacAddress());
@@ -3159,11 +3160,11 @@ public abstract class EnLinkdTestHelper {
                     assertEquals(1, shared.getBridgeIdsOnSegment().size());
                     assertTrue(shared.getBridgeIdsOnSegment().contains(nodeLId));
                     assertEquals(1, shared.getMacsOnSegment().size());
-                    assertEquals(1, SharedSegment.getBridgeMacLinks(shared).size());
-                    assertEquals(0, SharedSegment.getBridgeBridgeLinks(shared).size());
+                    assertEquals(1, TopologyService.getBridgeMacLinks(shared).size());
+                    assertEquals(0, TopologyService.getBridgeBridgeLinks(shared).size());
                     assertEquals(nodeLId.intValue(), shared.getDesignatedBridge().intValue());
                     assertEquals(portL5.intValue(), shared.getDesignatedPort().getBridgePort().intValue());
-                    BridgeMacLink link = SharedSegment.getBridgeMacLinks(shared).iterator().next();
+                    BridgeMacLink link = TopologyService.getBridgeMacLinks(shared).iterator().next();
                     assertEquals(nodeLId.intValue(), link.getNode().getId().intValue());
                     assertEquals(portL5,link.getBridgePort());
                     assertEquals(mac5,link.getMacAddress());
@@ -3172,11 +3173,11 @@ public abstract class EnLinkdTestHelper {
                     assertEquals(1, shared.getBridgeIdsOnSegment().size());
                     assertTrue(shared.getBridgeIdsOnSegment().contains(nodeLId));
                     assertEquals(1, shared.getMacsOnSegment().size());
-                    assertEquals(1, SharedSegment.getBridgeMacLinks(shared).size());
-                    assertEquals(0, SharedSegment.getBridgeBridgeLinks(shared).size());
+                    assertEquals(1, TopologyService.getBridgeMacLinks(shared).size());
+                    assertEquals(0, TopologyService.getBridgeBridgeLinks(shared).size());
                     assertEquals(nodeLId.intValue(), shared.getDesignatedBridge().intValue());
                     assertEquals(portL6.intValue(), shared.getDesignatedPort().getBridgePort().intValue());
-                    BridgeMacLink link = SharedSegment.getBridgeMacLinks(shared).iterator().next();
+                    BridgeMacLink link = TopologyService.getBridgeMacLinks(shared).iterator().next();
                     assertEquals(nodeLId.intValue(), link.getNode().getId().intValue());
                     assertEquals(portL6,link.getBridgePort());
                     assertEquals(mac6,link.getMacAddress());
@@ -3185,9 +3186,9 @@ public abstract class EnLinkdTestHelper {
                     assertEquals(1, shared.getBridgeIdsOnSegment().size());
                     assertTrue(shared.getBridgeIdsOnSegment().contains(nodeGId));
                     assertEquals(1,shared.getMacsOnSegment().size());
-                    assertEquals(1, SharedSegment.getBridgeMacLinks(shared).size());
-                    assertEquals(0, SharedSegment.getBridgeBridgeLinks(shared).size());
-                    BridgeMacLink link = SharedSegment.getBridgeMacLinks(shared).iterator().next();
+                    assertEquals(1, TopologyService.getBridgeMacLinks(shared).size());
+                    assertEquals(0, TopologyService.getBridgeBridgeLinks(shared).size());
+                    BridgeMacLink link = TopologyService.getBridgeMacLinks(shared).iterator().next();
                     assertEquals(nodeGId.intValue(), link.getNode().getId().intValue());
                     assertEquals(portG7.intValue(), link.getBridgePort().intValue());
                     assertEquals(mac7, link.getMacAddress());
@@ -3196,9 +3197,9 @@ public abstract class EnLinkdTestHelper {
                      assertEquals(1, shared.getBridgeIdsOnSegment().size());
                      assertTrue(shared.getBridgeIdsOnSegment().contains(nodeGId));
                      assertEquals(1,shared.getMacsOnSegment().size());
-                     assertEquals(1, SharedSegment.getBridgeMacLinks(shared).size());
-                     assertEquals(0, SharedSegment.getBridgeBridgeLinks(shared).size());
-                     BridgeMacLink link = SharedSegment.getBridgeMacLinks(shared).iterator().next();
+                     assertEquals(1, TopologyService.getBridgeMacLinks(shared).size());
+                     assertEquals(0, TopologyService.getBridgeBridgeLinks(shared).size());
+                     BridgeMacLink link = TopologyService.getBridgeMacLinks(shared).iterator().next();
                      assertEquals(nodeGId.intValue(), link.getNode().getId().intValue());
                      assertEquals(portG8.intValue(), link.getBridgePort().intValue());
                      assertEquals(mac8, link.getMacAddress());
@@ -3350,8 +3351,8 @@ public abstract class EnLinkdTestHelper {
         for (SharedSegment shared: shsegs) {
             assertFalse(shared.noMacsOnSegment());
             Set<Integer> nodeidsOnSegment = shared.getBridgeIdsOnSegment();
-            List<BridgeMacLink> links = SharedSegment.getBridgeMacLinks(shared);
-            List<BridgeBridgeLink> bblinks = SharedSegment.getBridgeBridgeLinks(shared);
+            List<BridgeMacLink> links = TopologyService.getBridgeMacLinks(shared);
+            List<BridgeBridgeLink> bblinks = TopologyService.getBridgeBridgeLinks(shared);
             if (shared.getMacsOnSegment().contains(macshared1)) {
                 assertEquals(2, nodeidsOnSegment.size());
                 assertEquals(3, shared.getMacsOnSegment().size());
@@ -3361,8 +3362,8 @@ public abstract class EnLinkdTestHelper {
                 assertTrue(nodeidsOnSegment.contains(nodeBId));
                 assertTrue(shared.getMacsOnSegment().contains(macshared2));
                 assertTrue(shared.getMacsOnSegment().contains(macshared3));
-                assertEquals(1, SharedSegment.getBridgeBridgeLinks(shared).size());
-                BridgeBridgeLink dlink = SharedSegment.getBridgeBridgeLinks(shared).iterator().next();
+                assertEquals(1, TopologyService.getBridgeBridgeLinks(shared).size());
+                BridgeBridgeLink dlink = TopologyService.getBridgeBridgeLinks(shared).iterator().next();
                 Set<String> macs = new HashSet<>();
                 if (revertedbblink) {
                 	assertEquals(nodeBId, dlink.getDesignatedNode().getId());
@@ -3398,7 +3399,7 @@ public abstract class EnLinkdTestHelper {
                 assertTrue(macs.contains(macshared3));
 
             } else if (shared.getMacsOnSegment().contains(macB21)) {
-                assertEquals(0, SharedSegment.getBridgeBridgeLinks(shared).size());
+                assertEquals(0, TopologyService.getBridgeBridgeLinks(shared).size());
                 assertEquals(1, nodeidsOnSegment.size());
                 assertTrue(nodeidsOnSegment.contains(nodeBId));
                 assertEquals(3, shared.getMacsOnSegment().size());
@@ -3417,7 +3418,7 @@ public abstract class EnLinkdTestHelper {
                assertTrue(macs.contains(macB22));
                assertTrue(macs.contains(macB23));
             } else if (shared.getMacsOnSegment().contains(macB71)) {
-                assertEquals(0, SharedSegment.getBridgeBridgeLinks(shared).size());
+                assertEquals(0, TopologyService.getBridgeBridgeLinks(shared).size());
                 assertEquals(1, nodeidsOnSegment.size());
                 assertTrue(nodeidsOnSegment.contains(nodeBId));
                 assertEquals(3, shared.getMacsOnSegment().size());
@@ -3436,7 +3437,7 @@ public abstract class EnLinkdTestHelper {
                assertTrue(macs.contains(macB72));
                assertTrue(macs.contains(macB73));
             } else if (shared.getMacsOnSegment().contains(macA11)) {
-                assertEquals(0, SharedSegment.getBridgeBridgeLinks(shared).size());
+                assertEquals(0, TopologyService.getBridgeBridgeLinks(shared).size());
                 assertEquals(1, nodeidsOnSegment.size());
                 assertTrue(nodeidsOnSegment.contains(nodeAId));
                 assertEquals(4, shared.getMacsOnSegment().size());
@@ -3457,7 +3458,7 @@ public abstract class EnLinkdTestHelper {
                assertTrue(macs.contains(macA13));
                assertTrue(macs.contains(macA14));
               } else if (shared.getMacsOnSegment().contains(macA61)) {
-                assertEquals(0, SharedSegment.getBridgeBridgeLinks(shared).size());
+                assertEquals(0, TopologyService.getBridgeBridgeLinks(shared).size());
                 assertEquals(1, nodeidsOnSegment.size());
                 assertTrue(nodeidsOnSegment.contains(nodeAId));
                 assertEquals(4, shared.getMacsOnSegment().size());
@@ -3525,8 +3526,8 @@ public abstract class EnLinkdTestHelper {
                 assertEquals(nodeAId,shared.getDesignatedBridge());
                 assertEquals(1, shared.getBridgeIdsOnSegment().size());
                 assertEquals(1, shared.getMacsOnSegment().size());
-                assertEquals(0, SharedSegment.getBridgeBridgeLinks(shared).size());
-                for (BridgeMacLink link: SharedSegment.getBridgeMacLinks(shared)) {
+                assertEquals(0, TopologyService.getBridgeBridgeLinks(shared).size());
+                for (BridgeMacLink link: TopologyService.getBridgeMacLinks(shared)) {
                     assertEquals(link.getBridgePort(),shared.getDesignatedPort().getBridgePort());
                     if (portA1.equals(link.getBridgePort())) {
                         assertEquals(mac1, link.getMacAddress());
@@ -3582,7 +3583,7 @@ public abstract class EnLinkdTestHelper {
             assertEquals(1, links.size());
             for (SharedSegment shared: links) {
                 assertFalse(shared.noMacsOnSegment());
-                assertEquals(0, SharedSegment.getBridgeBridgeLinks(shared).size());
+                assertEquals(0, TopologyService.getBridgeBridgeLinks(shared).size());
                 assertEquals(nodeAId,shared.getDesignatedBridge());
                 assertEquals(portA1,shared.getDesignatedPort().getBridgePort());
                 assertEquals(1, shared.getBridgeIdsOnSegment().size());
@@ -3592,7 +3593,7 @@ public abstract class EnLinkdTestHelper {
                 assertTrue(macs.contains(mac2));
                 assertTrue(macs.contains(mac3));
                 assertTrue(macs.contains(mac4));
-                for (BridgeMacLink link: SharedSegment.getBridgeMacLinks(shared)) {
+                for (BridgeMacLink link: TopologyService.getBridgeMacLinks(shared)) {
                     assertEquals(portA1, link.getBridgePort());
                 }
             }
@@ -3679,8 +3680,8 @@ public abstract class EnLinkdTestHelper {
                 assertFalse(shared.noMacsOnSegment());
                 assertEquals(nodeAId,shared.getDesignatedBridge());
                 assertEquals(1, shared.getBridgeIdsOnSegment().size());
-                assertEquals(0, SharedSegment.getBridgeBridgeLinks(shared).size());
-                for (BridgeMacLink link: SharedSegment.getBridgeMacLinks(shared)) {
+                assertEquals(0, TopologyService.getBridgeBridgeLinks(shared).size());
+                for (BridgeMacLink link: TopologyService.getBridgeMacLinks(shared)) {
                     assertEquals(link.getBridgePort(),shared.getDesignatedPort().getBridgePort());
                     if (portA1.equals(link.getBridgePort())) {
                         assertEquals(1, shared.getMacsOnSegment().size());
@@ -3799,9 +3800,9 @@ public abstract class EnLinkdTestHelper {
             assertEquals(10, shsegs.size());
             for (SharedSegment shared: shsegs) {
                 if (shared.noMacsOnSegment()) {
-                    assertEquals(0, SharedSegment.getBridgeMacLinks(shared).size());
-                    assertEquals(1, SharedSegment.getBridgeBridgeLinks(shared).size());
-                    BridgeBridgeLink link=SharedSegment.getBridgeBridgeLinks(shared).iterator().next();
+                    assertEquals(0, TopologyService.getBridgeMacLinks(shared).size());
+                    assertEquals(1, TopologyService.getBridgeBridgeLinks(shared).size());
+                    BridgeBridgeLink link= TopologyService.getBridgeBridgeLinks(shared).iterator().next();
                     if (reverse) {
                         assertEquals(nodeBId, link.getDesignatedNode().getId());
                         assertEquals(portBA,link.getDesignatedPort());
@@ -3816,7 +3817,7 @@ public abstract class EnLinkdTestHelper {
                     }
                 } else {
                     assertEquals(1, shared.getMacsOnSegment().size());
-                    BridgeMacLink link = SharedSegment.getBridgeMacLinks(shared).iterator().next();
+                    BridgeMacLink link = TopologyService.getBridgeMacLinks(shared).iterator().next();
                     if (link.getMacAddress().equals(mac1)) {
                         assertEquals(nodeAId, link.getNode().getId());
                         assertEquals(portA1,link.getBridgePort());
@@ -3926,8 +3927,8 @@ public abstract class EnLinkdTestHelper {
                 if (shared.getMacsOnSegment().contains(mac1)) {
                     assertEquals(nodeAId, shared.getDesignatedBridge());
                     assertEquals(portAB,shared.getDesignatedPort().getBridgePort());
-                    assertEquals(7, SharedSegment.getBridgeMacLinks(shared).size());
-                    assertEquals(1, SharedSegment.getBridgeBridgeLinks(shared).size());
+                    assertEquals(7, TopologyService.getBridgeMacLinks(shared).size());
+                    assertEquals(1, TopologyService.getBridgeBridgeLinks(shared).size());
                     assertEquals(2, shared.getBridgeIdsOnSegment().size());
                     assertEquals(7, shared.getMacsOnSegment().size());
                     assertTrue(shared.getMacsOnSegment().contains(mac2));
@@ -3938,33 +3939,33 @@ public abstract class EnLinkdTestHelper {
                     assertTrue(shared.getMacsOnSegment().contains(mac7));
                     assertFalse(shared.getMacsOnSegment().contains(mac8));
                     assertTrue(shared.getMacsOnSegment().contains(mac9));
-                    for (BridgeMacLink link: SharedSegment.getBridgeMacLinks(shared)) {
+                    for (BridgeMacLink link: TopologyService.getBridgeMacLinks(shared)) {
                         assertTrue(shared.getMacsOnSegment().contains(link.getMacAddress()));
                         assertEquals(link.getNode().getId().intValue(),nodeAId.intValue());
                         assertEquals(portAB,link.getBridgePort());
                     }
-                    BridgeBridgeLink dlink = SharedSegment.getBridgeBridgeLinks(shared).iterator().next();
+                    BridgeBridgeLink dlink = TopologyService.getBridgeBridgeLinks(shared).iterator().next();
                     assertEquals(nodeAId, dlink.getDesignatedNode().getId());
                     assertEquals(portAB, dlink.getDesignatedPort());
                     assertEquals(nodeBId, dlink.getNode().getId());
                     assertEquals(portBA, dlink.getBridgePort());
                 } else if (shared.getMacsOnSegment().contains(mac6)) {
-                    assertEquals(0, SharedSegment.getBridgeBridgeLinks(shared).size());
-                    assertEquals(1, SharedSegment.getBridgeMacLinks(shared).size());
+                    assertEquals(0, TopologyService.getBridgeBridgeLinks(shared).size());
+                    assertEquals(1, TopologyService.getBridgeMacLinks(shared).size());
                     assertEquals(1, shared.getMacsOnSegment().size());
                     assertEquals(nodeBId, shared.getDesignatedBridge());
                     assertEquals(portB6,shared.getDesignatedPort().getBridgePort());
-                    BridgeMacLink link = SharedSegment.getBridgeMacLinks(shared).iterator().next();
+                    BridgeMacLink link = TopologyService.getBridgeMacLinks(shared).iterator().next();
                     assertEquals(mac6, link.getMacAddress());
                     assertEquals(nodeBId,link.getNode().getId());
                     assertEquals(portB6,link.getBridgePort());
                 } else if (shared.getMacsOnSegment().contains(mac8)) {
-                    assertEquals(0, SharedSegment.getBridgeBridgeLinks(shared).size());
-                    assertEquals(1, SharedSegment.getBridgeMacLinks(shared).size());
+                    assertEquals(0, TopologyService.getBridgeBridgeLinks(shared).size());
+                    assertEquals(1, TopologyService.getBridgeMacLinks(shared).size());
                     assertEquals(1, shared.getMacsOnSegment().size());
                     assertEquals(nodeAId, shared.getDesignatedBridge());
                     assertEquals(portA8,shared.getDesignatedPort().getBridgePort());
-                    BridgeMacLink link = SharedSegment.getBridgeMacLinks(shared).iterator().next();
+                    BridgeMacLink link = TopologyService.getBridgeMacLinks(shared).iterator().next();
                     assertEquals(mac8, link.getMacAddress());
                     assertEquals(nodeAId,link.getNode().getId());
                     assertEquals(portA8,link.getBridgePort());
@@ -4038,10 +4039,10 @@ public abstract class EnLinkdTestHelper {
             for (SharedSegment shared: shsegs) {
                 assertFalse(shared.noMacsOnSegment());
                 if (shared.getMacsOnSegment().contains(macAB)) {
-                    assertEquals(1, SharedSegment.getBridgeBridgeLinks(shared).size());
+                    assertEquals(1, TopologyService.getBridgeBridgeLinks(shared).size());
                     assertEquals(1, shared.getMacsOnSegment().size());
                     assertEquals(2, shared.getBridgeIdsOnSegment().size());
-                    List<BridgeMacLink> links = SharedSegment.getBridgeMacLinks(shared);
+                    List<BridgeMacLink> links = TopologyService.getBridgeMacLinks(shared);
                     assertEquals(1, links.size());
                     for (BridgeMacLink link: links) {
                         assertEquals(macAB, link.getMacAddress());
@@ -4049,19 +4050,19 @@ public abstract class EnLinkdTestHelper {
                         assertEquals(portAB,link.getBridgePort());
                         assertEquals(BridgeMacLinkType.BRIDGE_LINK, link.getLinkType());
                     }
-                    BridgeBridgeLink dlink = SharedSegment.getBridgeBridgeLinks(shared).iterator().next();
+                    BridgeBridgeLink dlink = TopologyService.getBridgeBridgeLinks(shared).iterator().next();
                     assertEquals(nodeAId, dlink.getDesignatedNode().getId());
                     assertEquals(portAB, dlink.getDesignatedPort());
                     assertEquals(nodeBId, dlink.getNode().getId());
                     assertEquals(portBA, dlink.getBridgePort());
                 } else if (shared.getMacsOnSegment().contains(macA11)) {
-                    assertEquals(0, SharedSegment.getBridgeBridgeLinks(shared).size());
+                    assertEquals(0, TopologyService.getBridgeBridgeLinks(shared).size());
                     assertEquals(2, shared.getMacsOnSegment().size());
                     assertEquals(1, shared.getBridgeIdsOnSegment().size());
                     assertTrue(shared.getMacsOnSegment().contains(macA12));
                     assertEquals(nodeAId, shared.getDesignatedBridge());
                     assertEquals(portA1,shared.getDesignatedPort().getBridgePort());
-                    List<BridgeMacLink> links = SharedSegment.getBridgeMacLinks(shared);
+                    List<BridgeMacLink> links = TopologyService.getBridgeMacLinks(shared);
                     assertEquals(2, links.size());
                     for (BridgeMacLink link: links) {
                         assertEquals(nodeAId, link.getNode().getId());
@@ -4072,13 +4073,13 @@ public abstract class EnLinkdTestHelper {
                         }
                     }
                 } else if (shared.getMacsOnSegment().contains(macB21)) {
-                    assertEquals(0, SharedSegment.getBridgeBridgeLinks(shared).size());
+                    assertEquals(0, TopologyService.getBridgeBridgeLinks(shared).size());
                     assertEquals(2, shared.getMacsOnSegment().size());
                     assertEquals(1, shared.getBridgeIdsOnSegment().size());
                     assertTrue(shared.getMacsOnSegment().contains(macB22));
                     assertEquals(nodeBId, shared.getDesignatedBridge());
                     assertEquals(portB2,shared.getDesignatedPort().getBridgePort());
-                    List<BridgeMacLink> links = SharedSegment.getBridgeMacLinks(shared);
+                    List<BridgeMacLink> links = TopologyService.getBridgeMacLinks(shared);
                     assertEquals(2, links.size());
                     for (BridgeMacLink link: links) {
                         assertEquals(BridgeMacLinkType.BRIDGE_LINK, link.getLinkType());
@@ -4150,24 +4151,24 @@ public abstract class EnLinkdTestHelper {
 
             for (SharedSegment shared: shsegs) {
                 if (shared.getMacsOnSegment().size() == 0) {
-                    assertEquals(1, SharedSegment.getBridgeBridgeLinks(shared).size());
+                    assertEquals(1, TopologyService.getBridgeBridgeLinks(shared).size());
                     assertEquals(0, shared.getMacsOnSegment().size());
                     assertEquals(2, shared.getBridgeIdsOnSegment().size());
-                    List<BridgeMacLink> links = SharedSegment.getBridgeMacLinks(shared);
+                    List<BridgeMacLink> links = TopologyService.getBridgeMacLinks(shared);
                     assertEquals(0, links.size());
-                    BridgeBridgeLink dlink = SharedSegment.getBridgeBridgeLinks(shared).iterator().next();
+                    BridgeBridgeLink dlink = TopologyService.getBridgeBridgeLinks(shared).iterator().next();
                     assertEquals(nodeBId, dlink.getDesignatedNode().getId());
                     assertEquals(portBA, dlink.getDesignatedPort());
                     assertEquals(nodeAId, dlink.getNode().getId());
                     assertEquals(portAB, dlink.getBridgePort());
                 } else if (shared.getMacsOnSegment().contains(macA11)) {
-                    assertEquals(0, SharedSegment.getBridgeBridgeLinks(shared).size());
+                    assertEquals(0, TopologyService.getBridgeBridgeLinks(shared).size());
                     assertEquals(2, shared.getMacsOnSegment().size());
                     assertEquals(1, shared.getBridgeIdsOnSegment().size());
                     assertTrue(shared.getMacsOnSegment().contains(macA12));
                     assertEquals(nodeAId, shared.getDesignatedBridge());
                     assertEquals(portA1,shared.getDesignatedPort().getBridgePort());
-                    List<BridgeMacLink> links = SharedSegment.getBridgeMacLinks(shared);
+                    List<BridgeMacLink> links = TopologyService.getBridgeMacLinks(shared);
                     assertEquals(2, links.size());
                     for (BridgeMacLink link: links) {
                         assertEquals(nodeAId, link.getNode().getId());
@@ -4177,13 +4178,13 @@ public abstract class EnLinkdTestHelper {
                         }
                     }
                 } else if (shared.getMacsOnSegment().contains(macB21)) {
-                    assertEquals(0, SharedSegment.getBridgeBridgeLinks(shared).size());
+                    assertEquals(0, TopologyService.getBridgeBridgeLinks(shared).size());
                     assertEquals(2, shared.getMacsOnSegment().size());
                     assertEquals(1, shared.getBridgeIdsOnSegment().size());
                     assertTrue(shared.getMacsOnSegment().contains(macB22));
                     assertEquals(nodeBId, shared.getDesignatedBridge());
                     assertEquals(portB2,shared.getDesignatedPort().getBridgePort());
-                    List<BridgeMacLink> links = SharedSegment.getBridgeMacLinks(shared);
+                    List<BridgeMacLink> links = TopologyService.getBridgeMacLinks(shared);
                     assertEquals(2, links.size());
                     for (BridgeMacLink link: links) {
                         assertEquals(nodeBId, link.getNode().getId());
